Allow symforce codegen C++ databuffer types to change
This PR adds a C++ codegen config option for changing the type of the
databuffers generated during codegen. Before this change, the type of
all databuffers had to match the scalar type of the generated function.
E.g. the generated function would be templated on `Scalar`, and the
databuffers would be of type `Scalar*`. However, in some instances users
may have databuffers of a different scalar type. For example, if the
templated function is instantiated on `double` but the user wants to use
a databuffer of type `float`. This is especially useful in cases where
the underlying databuffer is large and a cast from e.g. `float` to
`double` may be expensive.

Note that this doesn't allow for the use of multiple databuffers of
different types. I'm not sure how common this is though, and it would be
somewhat complicated to support this, so I think simply allowing all
databuffers to use a different type is sufficient for 95% of use-cases.

Topic: databuffer_independent_scalar_type
GitOrigin-RevId: d4f26d743cdd60f93d154adc3679c4ae2da29f0b