[SymForce] Remove relative includes from sym
Bazel is unhappy with these on some compilers (others are fine), the
issues I've seen are that things build fine and are sandboxed correctly
but the reported includes from the compiler are absolute paths, which
trigger bazel's detection of undeclared includes.

I maybe have a bit of a concern that things can get weird if you e.g.
have sym as a subproject, and want to build against that, while sym is
also installed on your system or something (such that there are multiple
copies of sym available to the compiler).  This would basically not be a
concern with bazel, but could be with cmake.  Abseil does things this
way though, and is a dual CMake/Bazel project, so this is probably fine

Topic: sf-no-relative
Reviewers: brad,nathan,ryan-b
GitOrigin-RevId: 4a6afa3e0f1402a99b34377b3d7091df580edbb2