skymarshal: Turn format_field into show_field
Similarly to how store_field allows a pathspec giving the
destination of the field to store, allow format_field to also
take a path and print the json of the item at that path.

This also reworks both store and show to use int field indices
instead of strings, and introduces a new function translate_fields
that handles string to int conversion for both of them. This
allows for directly specifying indices to these methods in the
future without needing to go through strings at any point.

Because translate_fields is a static function, it ends up doing
some funky type magic to pass typing down through templated functions.

Also did a lot of other cleanup and simplification of the code.

Review by commit to exclude generated files.

GitOrigin-RevId: cfca8127c3b43dd6a46e44ce190cf08fc1f17d79