skymarshal: Support "omit_enum_wrapper" for #protobuf enums
Skymarshal currently defines any protobuf'd lcm enum as a full
empty message containing an enum named "Enum". This is annoying
in many ways, as it forces everything downstream to require .Enum or
_Enum in order to use the types. Python is even more annoying because
you can't easily alias nested classes on import.

Add an option that can be specified on new types to omit
this wrapper and produce a pure enum, as though it were declared
in native proto.

GitOrigin-RevId: 6a4713a4d514441d088e008d8b3e87d71ffd3791