Fix pybind stubs
This does a couple things:

- bump pybind11-stubgen to newest
- then patch things up so our stubs should be perfect on pybind (i.e.
  nothing missing)

The second one is just a matter of making sure things are available to
pybind at module definition time, i.e. the modules you depend on are
imported.

Making things automatically correct lets us delete a bunch of
post-hoc fixups in the stubs generation - the remaining one is
because of `tl::optional`

We have to do some more monkey patching of pybind11-stubgen - I think we
could upstream a good amount of that

Topic: pybind-stubs
GitOrigin-RevId: fa1655f2c82c197e9656a79798bba86ce3d961b0