[SymForce] Latex-compatible symbols
This adds an option to make symbol names format correctly as LaTeX,
which is _way_ nicer for notebooks.  Decided to make this optional
because outside of that use case, i.e. in plaintext, it is nicer to
have different formatting.  I've also updated the non-latex formatting
for Pose2 and Matrix to be better.

I also contemplated adding `\mathrm` around the `.`-separated names e.g.
in pose3, but that gets messy with things like dataclasses that use long
dot-separated chains. These do look ok just as math with `.` in them, so
I'm leaving them that way.

Topic: sf-latex-symbols
Reviewers: brad,nathan,ryan-b
GitOrigin-RevId: 74f3f98c73772c6027aa890ef0680794fc03b4da