[SymForce] Fix BAL example
The filtering introduced in
https://github.com/symforce-org/symforce/pull/390 had a couple issues:

1) It was computing the median as the point with the median norm.  This
   sorta works, it's not the median and could definitely behave weirdly
   sometimes.
2) The removal code renumbered the cameras without renumbering the
   factors.  We need to preserve the numbering (or create the factors
   with the correct numbering).

The bigger difference with other implementations though is that they
blindly project points behind the camera, while we were clamping the
projection into the half-space in front of the camera.  Clamping
produces very large residuals for the clamped points (because you end up
projecting something basically at infinity in the x-y plane).  Not
clamping produces something "not infinite" but not valid either (I left
a comment describing some of the reasoning here, and what you might do
instead in actual bundle adjustment problems).

I'm not currently going to dig into exactly what's going on with the BAL
problems, or whether these cameras/points make any sense or do anything
reasonable, but this allows the optimization to progress in the same way
that other solvers do.

On the largest dataset, we've also now revealed a new issue: we're using
the default SparseMatrix for the linearization, which uses `int` for
indices.  The largest BAL problem has more than 2^31 entries in the
hessian (at least in triplet form with duplicates), which causes
Eigen to wrap the index and segfault.  I'm not going to fix this for
now, but I've added an assert so that we detect this instead of
segfaulting.

Topic: sf-fix-bal
Reviewers: brad,ryan-b
Assignees: brad
GitOrigin-RevId: 8c730d46356d33b1027c62145884cff5aa92eb7b