Fix symforce codegen ValueError error message
We used `zip` on two lists known to be different sizes, which cuts off
the non-matching values in the error message. E.g. if
`formatted_symbols` has length 1 and `flattened_value` has length 0 we
would print `The following symbol/value pairs should match: []` which
isn't helpful.

This makes it so that we actually print the full list of symbols/values
that should match in the error message.

Topic: fix_codegen_util_error_msg
GitOrigin-RevId: dc1f6a946659cf33d813fb529409094e921d60be