Template LM solver on Values type
This PR refactors the LM solver such that datatypes other than a
`Values` object can be used with it. This should make it possible to
optimize a fixed-size state, which is useful in cases where the
structure of the optimization problem never changes, and thus we don't
need to pay the computational overhead associated with using `Values`
objects.

The end-goal is to be able to pass an object to the optimizer that
defines a `Retract` function as well as a linearization function which
takes the object as an argument. This should also greatly simplify the
construction of large linearization functions, since currently we
require linearization functions to have one argument per element in the
`Values`, which becomes unweildly for large problems (and uses excessive
CPU cycles to unpack the `Values` in order to actually call the
function).

I also added an extra TIC_TOC to measure the linearize time, since there
isn't one in the `Iterate()` function. There's one in the default
linearizer, but when users provide their own linearization function it
would be nice for it to be automatically instrumented.

Topic: values_agnostic_lm_solver
GitOrigin-RevId: 0328d79e9e5b2adfd1fb4018919093682871ddef