Add a fixed-size optimizer to symforce
This PR adds a new optimizer which is essentially a copy of the existing
symforce optimizer but without the use of `Values` objects. I opted to
only implement the minimum number of helper functions that I need for my
use-case in order to reduce code bloat; I figure we can add more helper
functions as they're needed instead of adding a bunch of unused
functions that we'll have to keep maintaining.

Topic: fixed_size_optimizer
Relative: optimize_static_func
GitOrigin-RevId: 875e6812eeaebb6778415f377b4b88cd2db8a004