[SymForce] Use uv for requirements
This lets us to an actual full solve for each python version, instead of
the ad-hoc patching we've been doing, which is much more scalable,
especially as parts of the ecosystem start dropping py3.8 and we're
trying to maintain py3.8 through 3.13 and beyond (at least for now?).
Afaik pip-tools isn't capable of doing a solve for python versions other
than what's installed, and uv is nicer anyway. uv also has this new
[lockfile](https://docs.astral.sh/uv/concepts/projects/layout/#the-lockfile)
thing they're pushing for, but it's uv-specific so I'd rather not lock us or
users into that.

Last I tried this, uv didn't support solving for python versions other
than the running interpreter; now they do (except that dependencies that
don't have wheels need to get built with an actual Python interpreter).

I previously had a GitHub action to build requirements for all Python
versions; it was super fast (<1m) and kinda convenient, it'd just make a
PR.  I think I like just having the requirements test do it now though,
now that it's possible to do that way.

Topic: sf-uv
Closes symforce-org/symforce#392
GitOrigin-RevId: 3e0ab436a7e50f501cea888f8e310fc39b382c5d