[SymForce] Add fast CreateIndex overload
`CreateIndex` currently takes a vector of Key, and does a bunch of map
lookups while making the index to find the entries for those keys.  This
is really silly, and extremely slow, when making an index for the whole
Values; in that case we'd rather just iterate directly, and sort after
if necessary.

Topic: sf-fast-index
GitOrigin-RevId: bbdcbe00097689247ca547f898349c39ae2dfa0b