[SymForce] Print unhandled type in ByTypeHelper
BY_TYPE_HELPER doesn't handle databuffers or preintegrated imu
measurements.  That's probably fine for now.  The case where I noticed
this was formatting a Values that contained a
PreintegratedImuMeasurements; formatting a Values like that is going to
generate a pretty huge string anyway that's maybe not super helpful, and
PreintegratedImuMeasurements doesn't have a formatter defined anyway.
Maybe a TODO to include those types here.

Topic: sf-bytypehelper-unhandled
Reviewers: ryan-b,nathan,xipeng
GitOrigin-RevId: 671a84d255eb4e90e7ec74f64c01cef956a4e691