[SymForce] No aligned storage
Using aligned storage for these breaks pulling them out of Values with
reinterpret_cast.  This is significantly faster than calling FromStorage

I don't think there's really a downside to this; I don't expect anything
to be doing vector ops on these types directly, and conversions between
the aligned and unaligned types are automatic

For raw Eigen types stored in Values, we still call FromStorage to
extract them.  I don't want to change the template parameter on `At`; we
could reinterpret_cast to the unaligned Eigen type and convert that to
the aligned one, but I'm not sure whether this matters

Topic: sf-no-align
Reviewers: matias,ryan-b,nathan,xipeng,jerry
GitOrigin-RevId: c47148ff683f7854c3d4293a0b3a6e8017d79580