[SymForce] Fix symforce-org/symforce#435
Numpy is super aggressive about converting things to numpy arrays when
calling `__mul__` on numpy scalars (or numpy arrays).

Our `geo.Matrix` gets converted because it has `__getitem__`.
SymEngine's `MutableDenseMatrix` gets converted because it has
`__array__`.

I can't think of a nice way around this.  So my approach is to go with a
rule "be very careful any time you left-multiply an unwrapped scalar
with something array-like that the result is the type you want."  I'm
not sure how practical this'll be in general.

Topic: sf-numpy-mul
GitOrigin-RevId: dd1f92b7625eeee5a3e3f5f4d787821dae564443