[SymForce] New Unit3
This PR seeks to revise the unit3 symforce geometry class. The previous
implementation stored S^2 representation as a `Rot3` that rotated the
Unit-Z vector across the 3D sphere manifold. This representation had
degenerate behavior in `retract` and `local_coordinates` in some cases.

 ## Updates

Following a previous symforce comment that was left, there are
alternative ways to representing `Unit3`'s underneath the hold. The new
implementation follows the structure outlined in Appendix B.2 of:

> [Hertzberg 2013] Integrating Generic Sensor Fusion Algorithms with
> Sound State Representations through Encapsulation of Manifolds.

With the new implementation `Unit3` now contains some additional
function calls:

- `from_unit_vector` - construction from unit `Vector3` without
  normalization check.
- `basis` - returning the basis of the tangent space in R^3.

This new `Unit3` change brings along with it some modifications to the
symforce `Group` and `LieGroup` structure. Because the `Unit3` manifold
is not a valid `Group` or `LieGroup`, certain tests and code generation
had to be updated to account for missing function calls.

`Unit3` no longer supports the operations:

- `identity`
- `inverse`
- `compose`
- `between`
- `to_tangent` - (just `local_coordinates` from identity element)
- `from_tangent` - (just `retract` from identity element)

All tests and codegen needed to be updated to reflect these changes.

`Unit3` class now requires an `epsilon` argument for `tangent_D_storage`
and `storage_D_tangent`. Because these functions are used in calculating
general jacbobians, all other calls to `tangent_D_storage` and
`storage_D_tangent` were updated to include the `epsilon` argument.

 ## Operation Count Increase

Most codegen functions that relied on the old `Unit3` implementation saw
a minor increase in the number of ops.

Topic: new-unit3
Reviewers: aaron, harvest, brad