[SymForce] Don't deallocate on ResetValues
Calling the assignment operator on `Values` is expensive, because it
means you deallocate everything the lhs contains (even if we called
the move assignment, and not the copy assignment, like we're doing
here).  And ditto for setting to `{}` and then letting it get set again.
This saves a bunch of time in the GncOptimizer which does this between
rounds of GNC (because mu has changed).

Topic: sf-no-realloc
Reviewers: nathan,ryan-b,xipeng
GitOrigin-RevId: 9d65df8c9288afd721a37ca17c7fee568aa7acf0