[SymForce] Don't use ternaries for index_entry_cache
This was resulting in `operator new(unsigned long)` being called (per
the generated assembly).  I _think_ this is because the type of the
ternary expression becomes a prvalue, per
[cppreference](https://en.cppreference.com/w/cpp/language/operator_other#Stage_4),
but I'm not 100% sure on this and I'm not sure how to confirm it.

This change eliminates the `operator new` on the happy path (we still of
course have to allocate if you don't pass an index_entry_cache), and
also eliminates time spent doing `__memmove_avx_unaligned_erms`
afterward which I _think_ is us copying the vector itself.

Topic: sf-ternary
Reviewers: nathan,ryan-b,xipeng
GitOrigin-RevId: bdd7035e933abd3bbe9be11c3b7e85e5fd3cd8d9