[SymForce] Maps
Copying from the stack-allocated fixed-size matrices to the
heap-allocated ones takes a noticeable amount of time for some problems.
This adds overloads for `Factor::Hessian` and `Factor::Jacobian` where
the generated function takes `Eigen::Map`s for the linearization, which
point to the heap-allocated matrices, to eliminate this.

Also adds the use_maps_for_outputs option to the CodegenConfig to
generate functions you can use for this.

Not swapping over all current generated factors to use this, although we
probably should, and should make this the default?

Topic: maps
Reviewers: ryan-b,nathan,xipeng
GitOrigin-RevId: d3081bac4f563f61fcf3785e088bb937ec42eed5