skymarshal: Check if array size can be represented in dimension size type
LCM dynamic sized arrays may specify a primitive type as the dimension size.
In that case, the maximum size of the array is determined by the largest
positive number that can be represented by the type. For instance,
`event_properties_t` contains dynamic sized arrays whose dimension
is `int8_t`. Hence, the maximum array length of these fields is 127.

skymarshal did not account for this so attempts to encode a message
with array length > 127 would succeed without warning. However, when
decoding the message, `std::length_error` would be thrown due to an
implicit conversion of the encoded size from `size_t` to the specified
dimension type (i.e., int8_t).

Skymarshal will now check if the length
of the array can be represented by the type (for dynamic LCM arrays). If
it exceeds the maximum length, it will fail to encode and return -1. The
below code is now generated:
```
+   if (this->bools.size() > static_cast<std::size_t>(std::numeric_limits<int8_t>::max())) return -1;
    int8_t v_num_bools = this->bools.size();
```

GitOrigin-RevId: 34315a13238a996711e4bedfc75999eac95c7f05