[SymForce] Support sympy-style signature for lambdify
sympy's lambdify accepts a list of inputs and an expression.  This is
easy to support, so why not, and it further reduces confusion relative
to the previous PR.

We could deduce the args from the output expressions, but at minimum
that won't give you the order you might have expected, and it seems
reasonable to match sympy's signature anyway.

Not supporting any additional arguments that sympy does:
https://docs.sympy.org/latest/modules/utilities/lambdify.html#sympy.utilities.lambdify.lambdify

`dummify` might be good, `modules` doesn't immediately translate well
(for single invocations, `math` is faster than `numpy` at least and
probably any of the alternatives, I'm not sure if `sympy.lambdify` can
produce broadcastable functions), `cse` is on by default, and
`docstring_limit` is just because sympy seems to have decided to render
the expression in the function docstring, which is usually a bad idea
for the reasons they state there and we don't do.

Topic: sf-lambdify-args
Relative: lambdify
Reviewers: andy,roy-v,nathan,ryan-br
GitOrigin-RevId: cdc2d733a74bf917a8eb9641adb2a45abd484f04