[SymForce] Fix build_wheels running on all export commits
This was running on every push to the export branch.  Instead it should
match the conditions we use for pull requests, i.e. run only while
not in draft state.

I'm also running this on all PRs, which will include the export PR, any
PRs exported for testing, and any PRs on the symforce repo.  I think
this is a good idea to catch bugs sooner, but if it's too much compute
we can dial it back.

Topic: sf-wheels-export
Reviewers: ryan-br,nathan

Random:
Assignees: ryan-br
GitOrigin-RevId: 386d9aa8ab44f64e36f067be7fc75953d07d4292