[SymForce] Loosen sympy version
We currently test on sympy 1.13, but upgrading to 1.13 only required
loosening things (see the changes to CppCodePrinter), so I think this is
ok.  Another reason we should test on our min requirements with uv.

1.14 isn't tested, and newer won't be tested, but that's probably ok.
We've noticed some issues on 1.14 where Adds are generated with
extraneous multiplies by `1.0` and such, and changes in distributivity
rules, but the former should be removed by the compiler and the latter
probably isn't a big deal.

Fixes symforce-org/symforce#449

Topic: sf-loose
GitOrigin-RevId: 28ae63ee07039c7df3c96542187a7374266bf430