[SymForce-External] Fix ruff format of types from wheel
Fix ruff format of types from wheel

format_py_dir doesn't work because it doesn't use the ruff config.

Example failure: https://github.com/symforce-org/symforce/actions/runs/20531137563/job/58982549124?pr=450

Here we have a line between 80 and 100 chars, so with the correct
config, it stays on the same line.  If you format it with an 80-char
config and then a 100-char config though, it's broken across lines and
the trailing comma is added, so formatting with 100 chars doesn't put it
back on one line.  And it has to be a list for the trailing comma thing
to happen, that's why this is pretty rare.

Topic: sf-ruff-wheel-types
Closes symforce-org/symforce#452
GitOrigin-RevId: 6b98c6c7f3a5ffc69c9ec471ac1a73de4ca01e7b